categories/web/web-easy-Emergency-access/give/identity_provider/requirements.txt
Flask==3.1.0
lxml==5.3.2
signxml==4.0.3
gunicorn==21.2.0
python-dotenv==1.0.1